PR and fixes
