+++ cellxgene_schema_cli/cellxgene_schema/cli.py
@@ -1,6 +1,6 @@
 import click
 import sys
-from cellxgene_schema import validate
+from .validate import validate
 
 
 @click.group(
